requirements.txt
kopf
pykube-ng
